Split fixtures in the Warehouse module
