+++ saleor/tests/fixtures.py
@@ -148,7 +148,6 @@ from ..site.models import SiteSettings
 from ..tax import TaxCalculationStrategy
 from ..tax.utils import calculate_tax_rate, get_tax_class_kwargs_for_order_line
 from ..thumbnail.models import Thumbnail
-from ..warehouse import WarehouseClickAndCollectOption
 from ..warehouse.models import (
     Allocation,
     PreorderAllocation,
@@ -7832,202 +7831,12 @@ def staff_notification_recipient(db, staff_user):
     return StaffNotificationRecipient.objects.create(active=True, user=staff_user)
 
 
-@pytest.fixture
-def warehouse(address, shipping_zone, channel_USD):
-    warehouse = Warehouse.objects.create(
-        address=address,
-        name="Example Warehouse",
-        slug="example-warehouse",
-        email="test@example.com",
-    )
-    warehouse.shipping_zones.add(shipping_zone)
-    warehouse.channels.add(channel_USD)
-    warehouse.save()
-    return warehouse
-
-
-@pytest.fixture
-def warehouse_with_external_ref(address, shipping_zone, channel_USD):
-    warehouse = Warehouse.objects.create(
-        address=address,
-        name="Example Warehouse With Ref",
-        slug="example-warehouse-with-ref",
-        email="test@example.com",
-        external_reference="example-warehouse-with-ref",
-    )
-    warehouse.shipping_zones.add(shipping_zone)
-    warehouse.channels.add(channel_USD)
-    warehouse.save()
-    return warehouse
-
-
-@pytest.fixture
-def warehouse_JPY(address, shipping_zone_JPY, channel_JPY):
-    warehouse = Warehouse.objects.create(
-        address=address,
-        name="Example Warehouse JPY",
-        slug="example-warehouse-jpy",
-        email="test-jpy@example.com",
-    )
-    warehouse.shipping_zones.add(shipping_zone_JPY)
-    warehouse.channels.add(channel_JPY)
-    warehouse.save()
-    return warehouse
-
-
-@pytest.fixture
-def warehouses(address, address_usa, channel_USD):
-    warehouses = Warehouse.objects.bulk_create(
-        [
-            Warehouse(
-                address=address.get_copy(),
-                name="Warehouse PL",
-                slug="warehouse1",
-                email="warehouse1@example.com",
-                external_reference="warehouse1",
-            ),
-            Warehouse(
-                address=address_usa.get_copy(),
-                name="Warehouse USA",
-                slug="warehouse2",
-                email="warehouse2@example.com",
-                external_reference="warehouse2",
-            ),
-        ]
-    )
-    for warehouse in warehouses:
-        warehouse.channels.add(channel_USD)
-    return warehouses
-
-
-@pytest.fixture
-def warehouses_for_cc(address, shipping_zones, channel_USD):
-    warehouses = Warehouse.objects.bulk_create(
-        [
-            Warehouse(
-                address=address.get_copy(),
-                name="Warehouse1",
-                slug="warehouse1",
-                email="warehouse1@example.com",
-            ),
-            Warehouse(
-                address=address.get_copy(),
-                name="Warehouse2",
-                slug="warehouse2",
-                email="warehouse2@example.com",
-                click_and_collect_option=WarehouseClickAndCollectOption.ALL_WAREHOUSES,
-            ),
-            Warehouse(
-                address=address.get_copy(),
-                name="Warehouse3",
-                slug="warehouse3",
-                email="warehouse3@example.com",
-                click_and_collect_option=WarehouseClickAndCollectOption.LOCAL_STOCK,
-                is_private=False,
-            ),
-            Warehouse(
-                address=address.get_copy(),
-                name="Warehouse4",
-                slug="warehouse4",
-                email="warehouse4@example.com",
-                click_and_collect_option=WarehouseClickAndCollectOption.LOCAL_STOCK,
-                is_private=False,
-            ),
-        ]
-    )
-    # add to shipping zones only not click and collect warehouses
-    warehouses[0].shipping_zones.add(*shipping_zones)
-    channel_USD.warehouses.add(*warehouses)
-    return warehouses
-
-
-@pytest.fixture
-def warehouse_for_cc(address, product_variant_list, channel_USD):
-    warehouse = Warehouse.objects.create(
-        address=address.get_copy(),
-        name="Local Warehouse",
-        slug="local-warehouse",
-        email="local@example.com",
-        is_private=False,
-        click_and_collect_option=WarehouseClickAndCollectOption.LOCAL_STOCK,
-    )
-    warehouse.channels.add(channel_USD)
-
-    Stock.objects.bulk_create(
-        [
-            Stock(
-                warehouse=warehouse, product_variant=product_variant_list[0], quantity=1
-            ),
-            Stock(
-                warehouse=warehouse, product_variant=product_variant_list[1], quantity=2
-            ),
-            Stock(
-                warehouse=warehouse, product_variant=product_variant_list[2], quantity=2
-            ),
-        ]
-    )
-    return warehouse
-
-
 @pytest.fixture(params=["warehouse_for_cc", "shipping_method"])
 def delivery_method(request, warehouse_for_cc, shipping_method):
     if request.param == "warehouse":
         return warehouse_for_cc
     if request.param == "shipping_method":
         return shipping_method
-
-
-@pytest.fixture
-def stocks_for_cc(warehouses_for_cc, product_variant_list, product_with_two_variants):
-    return Stock.objects.bulk_create(
-        [
-            Stock(
-                warehouse=warehouses_for_cc[0],
-                product_variant=product_variant_list[0],
-                quantity=5,
-            ),
-            Stock(
-                warehouse=warehouses_for_cc[1],
-                product_variant=product_variant_list[0],
-                quantity=3,
-            ),
-            Stock(
-                warehouse=warehouses_for_cc[1],
-                product_variant=product_variant_list[1],
-                quantity=10,
-            ),
-            Stock(
-                warehouse=warehouses_for_cc[1],
-                product_variant=product_variant_list[2],
-                quantity=10,
-            ),
-            Stock(
-                warehouse=warehouses_for_cc[2],
-                product_variant=product_variant_list[0],
-                quantity=3,
-            ),
-            Stock(
-                warehouse=warehouses_for_cc[3],
-                product_variant=product_variant_list[0],
-                quantity=3,
-            ),
-            Stock(
-                warehouse=warehouses_for_cc[3],
-                product_variant=product_variant_list[1],
-                quantity=3,
-            ),
-            Stock(
-                warehouse=warehouses_for_cc[3],
-                product_variant=product_with_two_variants.variants.last(),
-                quantity=7,
-            ),
-            Stock(
-                warehouse=warehouses_for_cc[3],
-                product_variant=product_variant_list[2],
-                quantity=3,
-            ),
-        ]
-    )
 
 
 @pytest.fixture
@@ -8162,139 +7971,6 @@ def checkout_with_prices(
     user.save(update_fields=["metadata"])
 
     return checkout_with_items
-
-
-@pytest.fixture
-def warehouses_with_shipping_zone(warehouses, shipping_zone):
-    warehouses[0].shipping_zones.add(shipping_zone)
-    warehouses[1].shipping_zones.add(shipping_zone)
-    return warehouses
-
-
-@pytest.fixture
-def warehouses_with_different_shipping_zone(warehouses, shipping_zones):
-    warehouses[0].shipping_zones.add(shipping_zones[0])
-    warehouses[1].shipping_zones.add(shipping_zones[1])
-    return warehouses
-
-
-@pytest.fixture
-def warehouse_no_shipping_zone(address, channel_USD):
-    warehouse = Warehouse.objects.create(
-        address=address,
-        name="Warehouse without shipping zone",
-        slug="warehouse-no-shipping-zone",
-        email="test2@example.com",
-        external_reference="warehouse-no-shipping-zone",
-    )
-    warehouse.channels.add(channel_USD)
-    return warehouse
-
-
-@pytest.fixture
-def stock(variant, warehouse):
-    return Stock.objects.create(
-        product_variant=variant, warehouse=warehouse, quantity=15
-    )
-
-
-@pytest.fixture
-def allocation(order_line, stock):
-    return Allocation.objects.create(
-        order_line=order_line, stock=stock, quantity_allocated=order_line.quantity
-    )
-
-
-@pytest.fixture
-def allocations(order_list, stock, channel_USD):
-    variant = stock.product_variant
-    product = variant.product
-    channel_listing = variant.channel_listings.get(channel=channel_USD)
-    net = variant.get_price(channel_listing)
-    gross = Money(amount=net.amount * Decimal(1.23), currency=net.currency)
-    price = TaxedMoney(net=net, gross=gross)
-    lines = OrderLine.objects.bulk_create(
-        [
-            OrderLine(
-                order=order_list[0],
-                variant=variant,
-                quantity=1,
-                product_name=str(variant.product),
-                variant_name=str(variant),
-                product_sku=variant.sku,
-                product_variant_id=variant.get_global_id(),
-                is_shipping_required=variant.is_shipping_required(),
-                is_gift_card=variant.is_gift_card(),
-                unit_price=price,
-                total_price=price,
-                tax_rate=Decimal("0.23"),
-                **get_tax_class_kwargs_for_order_line(product.product_type.tax_class),
-            ),
-            OrderLine(
-                order=order_list[1],
-                variant=variant,
-                quantity=2,
-                product_name=str(variant.product),
-                variant_name=str(variant),
-                product_sku=variant.sku,
-                product_variant_id=variant.get_global_id(),
-                is_shipping_required=variant.is_shipping_required(),
-                is_gift_card=variant.is_gift_card(),
-                unit_price=price,
-                total_price=price,
-                tax_rate=Decimal("0.23"),
-                **get_tax_class_kwargs_for_order_line(product.product_type.tax_class),
-            ),
-            OrderLine(
-                order=order_list[2],
-                variant=variant,
-                quantity=4,
-                product_name=str(variant.product),
-                variant_name=str(variant),
-                product_sku=variant.sku,
-                product_variant_id=variant.get_global_id(),
-                is_shipping_required=variant.is_shipping_required(),
-                is_gift_card=variant.is_gift_card(),
-                unit_price=price,
-                total_price=price,
-                tax_rate=Decimal("0.23"),
-                **get_tax_class_kwargs_for_order_line(product.product_type.tax_class),
-            ),
-        ]
-    )
-
-    for order in order_list:
-        order.search_vector = FlatConcatSearchVector(
-            *prepare_order_search_vector_value(order)
-        )
-    Order.objects.bulk_update(order_list, ["search_vector"])
-
-    return Allocation.objects.bulk_create(
-        [
-            Allocation(
-                order_line=lines[0], stock=stock, quantity_allocated=lines[0].quantity
-            ),
-            Allocation(
-                order_line=lines[1], stock=stock, quantity_allocated=lines[1].quantity
-            ),
-            Allocation(
-                order_line=lines[2], stock=stock, quantity_allocated=lines[2].quantity
-            ),
-        ]
-    )
-
-
-@pytest.fixture
-def preorder_allocation(
-    order_line, preorder_variant_global_and_channel_threshold, channel_PLN
-):
-    variant = preorder_variant_global_and_channel_threshold
-    product_variant_channel_listing = variant.channel_listings.get(channel=channel_PLN)
-    return PreorderAllocation.objects.create(
-        order_line=order_line,
-        product_variant_channel_listing=product_variant_channel_listing,
-        quantity=order_line.quantity,
-    )
 
 
 @pytest.fixture
