+++ saleor/warehouse/tests/fixtures/__init__.py
@@ -1,0 +1,4 @@
+from .allocation import *  # noqa: F403
+from .preorder_allocation import *  # noqa: F403
+from .stock import *  # noqa: F403
+from .warehouse import *  # noqa: F403
